chore(team): split remote relay types
